Refactored utilities to feature_importance_utils.py
