+++ compiler_opt/tools/feature_importance.py
@@ -49,193 +49,27 @@ from compiler_opt.rl import data_reader
 from compiler_opt.rl import constant
 from compiler_opt.rl import registry
 
+from compiler_opt.tools import feature_importance_utils
+
 import tensorflow as tf
 import shap
 import numpy
 import numpy.typing
 import json
 
-from tf_agents.typing import types
-from typing import Callable, Dict, Tuple
-
-SignatureType = Dict[str, Tuple[numpy.typing.ArrayLike, tf.dtypes.DType]]
-
-# only create flags if running as a script to prevent interference with pytest
-if __name__ == '__main__':
-  _DATA_PATH = flags.DEFINE_multi_string(
-      'data_path', [], 'Path to TFRecord file(s) containing trace data.')
-  _MODEL_PATH = flags.DEFINE_string('model_path', '',
-                                    'Path to the model to explain')
-  _OUTPUT_FILE = flags.DEFINE_string(
-      'output_file', '',
-      'The path to the output file containing the SHAP values')
-  _NUM_EXAMPLES = flags.DEFINE_integer(
-      'num_examples', 1, 'The number of examples to process from the trace')
-  _GIN_FILES = flags.DEFINE_multi_string(
-      'gin_files', [], 'List of paths to gin configuration files.')
-  _GIN_BINDINGS = flags.DEFINE_multi_string(
-      'gin_bindings', [],
-      'Gin bindings to override the values set in the config files.')
-
-
-def get_input_signature(example_input: types.NestedTensorSpec) -> SignatureType:
-  """Gets the signature of an observation
-
-  This function takes in an example input and returns a signature of that
-  input containing all of the info needed to restructure a flat array back into
-  the original format later on. This function returns a dictionary with the
-  same keys as the original input but with the items being tuples where the
-  first value is the shape of that feature and the second is its data type.
-
-  Args:
-    example_input: a nested tensor spec (dictionary of tensors) that serves
-      as an example for generating the signature.
-  """
-  input_signature = {}
-  for input_key in example_input:
-    input_signature[input_key] = (tf.shape(example_input[input_key]).numpy(),
-                                  example_input[input_key].dtype)
-  return input_signature
-
-
-def get_signature_total_size(input_signature: SignatureType) -> int:
-  """Gets the total number of elements in a single problem instance
-
-  Args:
-    input_signature: An input signature to calculate the number of elements in
-  """
-  total_size = 0
-  for input_key in input_signature:
-    total_size += numpy.prod(input_signature[input_key][0])
-  return total_size
-
-
-def pack_flat_array_into_input(
-    flat_array: numpy.typing.ArrayLike,
-    signature_spec: SignatureType) -> types.NestedTensorSpec:
-  """Packs a flat array into a nested tensor spec to feed into a model
-
-  Args:
-    flat_array: The data to be packed back into the specified nested tensor
-      specification
-    signature_spec: A signature that is used to create the correct structure
-      for all of the values in the flat array
-  """
-  output_input_dict = {}
-  current_index = 0
-  for needed_input in signature_spec:
-    part_size = numpy.prod(signature_spec[needed_input][0])
-    needed_subset = flat_array[current_index:current_index + part_size]
-    current_index += part_size
-    output_input_dict[needed_input] = tf.cast(
-        tf.constant(needed_subset, shape=signature_spec[needed_input][0]),
-        dtype=signature_spec[needed_input][1])
-  return output_input_dict
-
-
-def flatten_input(to_flatten: types.NestedTensorSpec,
-                  array_size: int) -> numpy.typing.ArrayLike:
-  """Flattens problem instance data into a flat array for shap
-
-  Args:
-    to_flatten: A nested tensor spec of data that needs to be flattend into
-      an array
-    array_size: An integer representing the size of the output array. Used for
-      allocating the flat array to place all the data in.
-  """
-  output_array = numpy.empty(array_size)
-  input_index = 0
-  for input_key in to_flatten:
-    current_size = tf.size(to_flatten[input_key])
-    end_index = input_index + current_size
-    output_array[input_index:end_index] = to_flatten[input_key].numpy().astype(
-        numpy.float32)
-    input_index += current_size
-  return output_array
-
-
-def process_raw_trajectory(
-    raw_trajectory: types.ForwardRef) -> types.NestedTensorSpec:
-  """Processes the raw example data into a nested tensor spec that can be
-  easily fed into a model.
-
-  Args:
-    raw_trajectory: Raw data representing an individual problem instance from
-      a trace.
-  """
-  observation = raw_trajectory.observation
-  observation.update({
-      'step_type': raw_trajectory.step_type,
-      'reward': raw_trajectory.reward,
-      'discount': raw_trajectory.discount
-  })
-
-  # remove batch size dimension
-  for key in observation:
-    observation[key] = tf.squeeze(observation[key], axis=0)
-
-  return observation
-
-
-def collapse_values(
-    input_signature: SignatureType,
-    shap_values: numpy.typing.ArrayLike) -> numpy.typing.ArrayLike:
-  """Collapses shap values so that there is only a single value per feature
-
-  Args:
-    input_signature: The signature of the model input. Used to determine what
-      (if any) features need to be collapsed.
-    shap_values: A numpy array of shap values that need to be processed.
-  """
-  output_shap_values = numpy.empty((_NUM_EXAMPLES.value, len(input_signature)))
-  for i in range(0, _NUM_EXAMPLES.value):
-    current_index = 0
-    current_feature = 0
-    for input_key in input_signature:
-      part_size = numpy.prod(input_signature[input_key][0])
-      output_shap_values[i, current_feature] = numpy.sum(
-          shap_values[i, current_index:current_index + part_size])
-      current_feature += 1
-      current_index += part_size
-  return output_shap_values
-
-
-def get_max_part_size(input_signature: SignatureType) -> int:
-  """Gets the size (as a single scalar) of the largest feature in terms of
-  the number of elements.
-
-  Args:
-    input_signature: The input signature that we want to find the largest
-      feature in.
-  """
-  part_sizes = numpy.empty(len(input_signature))
-  for index, input_key in enumerate(input_signature):
-    part_sizes[index] = numpy.prod(input_signature[input_key][0])
-  return numpy.max(part_sizes)
-
-
-def create_run_model_function(action_fn: Callable,
-                              input_sig: SignatureType) -> Callable:
-  """Returns a function that takes in a flattend input array and returns the
-  model output as a scalar.
-
-  Args:
-    action_fn: The action function from the tensorflow saved model saved
-      through tf_agents
-    input_sig: The input signature for the model currently under analysis.
-      Used to pack the flat array back into a nested tensor spec.
-  """
-
-  def run_model(flat_input_array):
-    output = numpy.empty(flat_input_array.shape[0])
-    for index, flat_input in enumerate(flat_input_array):
-      input_dict = pack_flat_array_into_input(flat_input, input_sig)
-      model_output = action_fn(**input_dict).items()
-      # get the value of the first item as a numpy array
-      output[index] = list(model_output)[0][1].numpy()[0]
-    return output
-
-  return run_model
+_DATA_PATH = flags.DEFINE_multi_string(
+    'data_path', [], 'Path to TFRecord file(s) containing trace data.')
+_MODEL_PATH = flags.DEFINE_string('model_path', '',
+                                  'Path to the model to explain')
+_OUTPUT_FILE = flags.DEFINE_string(
+    'output_file', '', 'The path to the output file containing the SHAP values')
+_NUM_EXAMPLES = flags.DEFINE_integer(
+    'num_examples', 1, 'The number of examples to process from the trace')
+_GIN_FILES = flags.DEFINE_multi_string(
+    'gin_files', [], 'List of paths to gin configuration files.')
+_GIN_BINDINGS = flags.DEFINE_multi_string(
+    'gin_bindings', [],
+    'Gin bindings to override the values set in the config files.')
 
 
 def main(_):
@@ -260,29 +94,33 @@ def main(_):
   saved_policy = tf.saved_model.load(_MODEL_PATH.value)
   action_fn = saved_policy.signatures['action']
 
-  observation = process_raw_trajectory(raw_trajectory)
-  input_sig = get_input_signature(observation)
+  observation = feature_importance_utils.process_raw_trajectory(raw_trajectory)
+  input_sig = feature_importance_utils.get_input_signature(observation)
 
-  run_model = create_run_model_function(action_fn, input_sig)
+  run_model = feature_importance_utils.create_run_model_function(
+      action_fn, input_sig)
 
-  total_size = get_signature_total_size(input_sig)
-  flattened_input = flatten_input(observation, total_size)
+  total_size = feature_importance_utils.get_signature_total_size(input_sig)
+  flattened_input = feature_importance_utils.flatten_input(
+      observation, total_size)
   flattened_input = numpy.expand_dims(flattened_input, axis=0)
   dataset = numpy.empty((_NUM_EXAMPLES.value, total_size))
   for i in range(0, _NUM_EXAMPLES.value):
     raw_trajectory = next(dataset_iter)
-    observation = process_raw_trajectory(raw_trajectory)
-    flat_input = flatten_input(observation, total_size)
+    observation = feature_importance_utils.process_raw_trajectory(
+        raw_trajectory)
+    flat_input = feature_importance_utils.flatten_input(observation, total_size)
     dataset[i] = flat_input
 
   explainer = shap.KernelExplainer(run_model, numpy.zeros((1, total_size)))
   shap_values = explainer.shap_values(dataset, nsamples=1000)
-  processed_shap_values = collapse_values(input_sig, shap_values)
+  processed_shap_values = feature_importance_utils.collapse_values(
+      input_sig, shap_values, _NUM_EXAMPLES.value)
 
   # if we have more than one value per feature, just set the dataset to zeros
   # as summing across a dimension produces data that doesn't really mean
   # anything
-  if get_max_part_size(input_sig) > 1:
+  if feature_importance_utils.get_max_part_size(input_sig) > 1:
     dataset = numpy.zeros(processed_shap_values.shape)
 
   feature_names = list(input_sig.keys())
